refactor: split runtime types into separate files
Just makes it a bit easier to navigate.
