+++ common/schema/databaseruntime.go
@@ -9,9 +9,9 @@ type DatabaseRuntime struct {
 	Connections *DatabaseRuntimeConnections `parser:"" protobuf:"1,optional"`
 }
 
-func (m *DatabaseRuntime) runtimeElement() {
+var _ Runtime = (*DatabaseRuntime)(nil)
 
-}
+func (m *DatabaseRuntime) runtimeElement() {}
 
 var _ Symbol = (*DatabaseRuntime)(nil)
