+++ common/schema/moduleruntime.go
@@ -1,13 +1,5 @@
 package schema
 
-import (
-	"time"
-
-	"github.com/alecthomas/types/optional"
-
-	"github.com/block/ftl/internal/key"
-)
-
 type DeploymentState int
 
 const (
@@ -28,13 +20,6 @@ type Runtime interface {
 	runtimeElement()
 }
 
-var _ Runtime = (*ModuleRuntimeScaling)(nil)
-var _ Runtime = (*ModuleRuntimeRunner)(nil)
-var _ Runtime = (*ModuleRuntimeDeployment)(nil)
-var _ Runtime = (*VerbRuntime)(nil)
-var _ Runtime = (*TopicRuntime)(nil)
-var _ Runtime = (*DatabaseRuntime)(nil)
-
 // ModuleRuntime is runtime configuration for a module that can be dynamically updated.
 type ModuleRuntime struct {
 	Base       ModuleRuntimeBase        `protobuf:"1"` // Base is always present.
@@ -43,65 +28,11 @@ type ModuleRuntime struct {
 	Runner     *ModuleRuntimeRunner     `protobuf:"4,optional"`
 }
 
-type ModuleRuntimeBase struct {
-	CreateTime time.Time `protobuf:"1"`
-	Language   string    `protobuf:"2"`
-	OS         string    `protobuf:"3,optional"`
-	Arch       string    `protobuf:"4,optional"`
-	// Image is the name of the runner image. Defaults to "ftl0/ftl-runner".
-	// Must not include a tag, as FTL's version will be used as the tag.
-	Image string `protobuf:"5,optional"`
-}
-
-//protobuf:1
-type ModuleRuntimeDeployment struct {
-	DeploymentKey key.Deployment             `protobuf:"2"`
-	CreatedAt     time.Time                  `protobuf:"3"`
-	ActivatedAt   optional.Option[time.Time] `protobuf:"4"`
-	State         DeploymentState            `protobuf:"5"`
-}
-
-func (m *ModuleRuntimeDeployment) runtimeElement() {
-
-}
-
-//protobuf:2
-type ModuleRuntimeScaling struct {
-	MinReplicas int32 `protobuf:"1"`
-}
-
-func (m *ModuleRuntimeScaling) runtimeElement() {
-}
-
-//protobuf:3
-type ModuleRuntimeRunner struct {
-	// Endpoint is the endpoint of the deployed module.
-	Endpoint string `protobuf:"1"`
-}
-
-func (m *ModuleRuntimeRunner) runtimeElement() {
-}
-
-func (m *ModuleRuntimeRunner) GetEndpoint() string {
-	if m == nil {
-		return ""
-	}
-	return m.Endpoint
-
-}
-
 func (m *ModuleRuntime) GetScaling() *ModuleRuntimeScaling {
 	if m == nil {
 		return nil
 	}
 	return m.Scaling
-}
-
-func (m *ModuleRuntimeScaling) GetMinReplicas() int32 {
-	if m == nil {
-		return 0
-	}
-	return m.MinReplicas
 }
 
 func (m *ModuleRuntime) GetDeployment() *ModuleRuntimeDeployment {
@@ -128,20 +59,6 @@ func (m *ModuleRuntime) ModDeployment() *ModuleRuntimeDeployment {
 		m.Deployment = &ModuleRuntimeDeployment{}
 	}
 	return m.Deployment
-}
-
-func (m *ModuleRuntimeDeployment) GetCreatedAt() time.Time {
-	if m == nil {
-		return time.Time{}
-	}
-	return m.CreatedAt
-}
-
-func (m *ModuleRuntimeDeployment) GetDeploymentKey() key.Deployment {
-	if m == nil {
-		return key.Deployment{}
-	}
-	return m.DeploymentKey
 }
 
 func (m *ModuleRuntime) ModScaling() *ModuleRuntimeScaling {
